# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HZ Navigator API",
    "version": "1.0.0",
    "description": "# HZ Navigator API Documentation\n\nHZ Navigator is a comprehensive HUBZone certification management platform that helps businesses maintain compliance with SBA HUBZone program requirements.\n\n## Features\n\n- **Authentication**: JWT-based authentication with role-based access control\n- **Business Management**: Create and manage HUBZone certified businesses\n- **Compliance Monitoring**: Real-time compliance tracking and alerts\n- **Document Management**: Secure document upload and OCR processing\n- **Contract Tracking**: Agency contract management and goal tracking\n- **HUBZone Maps**: Geographic verification with vector tile support\n- **Analytics**: Comprehensive reporting and analytics dashboards\n\n## Authentication\n\nMost endpoints require authentication via JWT Bearer token:\n\n```\nAuthorization: Bearer <your-jwt-token>\n```\n\n## Rate Limiting\n\n- Authentication endpoints: 5 requests per minute\n- Password reset: 3 requests per hour\n- Standard API: 100 requests per minute\n- File uploads: 10 requests per minute\n- Admin endpoints: 30 requests per minute",
    "contact": {
      "name": "HZ Navigator Support",
      "email": "support@hz-navigator.gov",
      "url": "https://hz-navigator.gov/support"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://hz-navigator.gov/terms"
    }
  },
  "servers": [
    {
...
    },
    {
...
    },
    {
...
  "paths": {
...
    "/api/hubzones": {
      "get": {
        "summary": "List all HUBZone areas",
        "tags": ["HUBZones"],
        "security": [],
        "parameters": [
          { "$ref": "#/components/parameters/PageParam" },
          { "$ref": "#/components/parameters/LimitParam" },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "List of HUBZone areas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Hubzone" } },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/hubzones/check": {
      "post": {
        "summary": "Check if coordinates are in a HUBZone",
        "tags": ["HUBZones"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["latitude", "longitude"],
                "properties": {
                  "latitude": { "type": "number", "minimum": -90, "maximum": 90 },
                  "longitude": { "type": "number", "minimum": -180, "maximum": 180 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "HUBZone check result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HubzoneCheckResult" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      },
...
      "Hubzone": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "geoid": { "type": "string" },
          "name": { "type": "string" },
          "zoneType": { "type": "string" },
          "status": { "type": "string", "enum": ["active", "expired", "pending", "redesignated"] },
          "state": { "type": "string" },
          "county": { "type": "string" }
        }
      },
      "HubzoneCheckResult": {
        "type": "object",
        "properties": {
          "isInHubzone": { "type": "boolean" },
          "hubzones": { "type": "array", "items": { "$ref": "#/components/schemas/Hubzone" } }
        }
      },
...
    },
...
    "parameters": {
      "PageParam": {
        "name": "page",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "default": 1 }
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
      },
...
    }
  },
  "security": [{ "BearerAuth": [] }]
}